Add command to send private messages to users
adds capability of send direct messages to one specific user.
This message is treated as a /secret and will not be shown to others users in
the same room.
Changes were made in the client and server. The protocol must be respect.
Another architectures changes must be made probably soon.